move checkbox and textInput components to their own files cleanup margins padding on fellowship
