+++ packages/web/src/alliance/SignupForm.tsx
@@ -1,8 +1,9 @@
 import * as React from 'react'
 import { StyleSheet, View } from 'react-native'
 import { NewMember } from 'src/alliance/AllianceMember'
+import { CheckboxWithLabel } from 'src/forms/CheckboxWithLabel'
 import FormContainer, { emailIsValid, hasField } from 'src/forms/Form'
-import { CheckboxWithLabel, Form } from 'src/forms/FormComponents'
+import { Form } from 'src/forms/FormComponents'
 import { LabeledInput } from 'src/forms/LabeledInput'
 import { NameSpaces, useTranslation } from 'src/i18n'
 import Checkmark from 'src/icons/Checkmark'
